refactor: split download manager support modules
