+++ src-tauri/src/core/manager.rs
@@ -3,6 +3,11 @@
 //! This module provides the main DownloadManager that orchestrates all download operations,
 //! manages concurrent downloads, and handles progress tracking and event emission.
 
+mod events;
+mod integrity;
+mod queue;
+mod stats;
+
 use sha2::{Digest, Sha256};
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
@@ -515,70 +520,6 @@ impl DownloadManager {
         fs::write(&self.state_path, json)
             .await
             .map_err(|e| AppError::System(format!("Failed to write persisted state: {}", e)))
-    }
-
-    fn recompute_stats(&mut self) {
-        let total_tasks = self.tasks.len();
-        let completed_tasks = self
-            .tasks
-            .values()
-            .filter(|t| t.status == TaskStatus::Completed)
-            .count();
-        let failed_tasks = self
-            .tasks
-            .values()
-            .filter(|t| t.status == TaskStatus::Failed)
-            .count();
-        let active_downloads = self.active_downloads.len();
-
-        let total_downloaded: u64 = self
-            .tasks
-            .values()
-            .filter(|t| t.status == TaskStatus::Completed)
-            .map(|t| t.downloaded_size)
-            .sum();
-
-        let current_speeds: Vec<f64> = self
-            .tasks
-            .values()
-            .filter(|t| t.status == TaskStatus::Downloading)
-            .map(|t| t.speed)
-            .collect();
-
-        let average_speed = if current_speeds.is_empty() {
-            0.0
-        } else {
-            current_speeds.iter().sum::<f64>() / current_speeds.len() as f64
-        };
-        let display_total_speed_bps = self
-            .tasks
-            .values()
-            .filter(|t| t.status == TaskStatus::Downloading)
-            .map(|t| t.display_speed_bps)
-            .sum();
-        let average_transfer_duration =
-            Self::average_metric(&self.lifecycle_metrics.transfer_duration_secs);
-        let average_commit_duration =
-            Self::average_metric(&self.lifecycle_metrics.commit_duration_secs);
-        let p95_commit_duration =
-            Self::percentile_metric(&self.lifecycle_metrics.commit_duration_secs, 0.95);
-
-        self.stats = ModelsDownloadStats {
-            total_tasks,
-            completed_tasks,
-            failed_tasks,
-            total_downloaded,
-            average_speed,
-            display_total_speed_bps,
-            active_downloads,
-            queue_paused: self.queue_paused,
-            average_transfer_duration,
-            average_commit_duration,
-            p95_commit_duration,
-            failed_commit_count: self.lifecycle_metrics.failed_commit_count,
-            commit_warning_count: self.lifecycle_metrics.commit_warning_count,
-            commit_elevated_warning_count: self.lifecycle_metrics.commit_elevated_warning_count,
-        };
     }
 
     /// Start the download manager
@@ -2008,137 +1949,6 @@ impl DownloadManager {
         resumed
     }
 
-    async fn enqueue_task(&self, task_id: &str, priority: u8) -> bool {
-        let inserted = {
-            let mut queue = self.task_queue.lock().await;
-            if queue.iter().any(|item| item.task_id == task_id) {
-                return false;
-            }
-
-            queue.push(TaskPriority {
-                task_id: task_id.to_string(),
-                priority,
-                created_at: chrono::Utc::now(),
-            });
-            true
-        };
-
-        if inserted {
-            if let Err(err) = self.persist_state().await {
-                warn!("Failed to persist state after enqueue: {}", err);
-            }
-        }
-        inserted
-    }
-
-    async fn remove_task_from_queue(&self, task_id: &str) -> bool {
-        let removed = {
-            let mut queue = self.task_queue.lock().await;
-            if queue.is_empty() {
-                return false;
-            }
-
-            let mut items = Vec::with_capacity(queue.len());
-            while let Some(item) = queue.pop() {
-                items.push(item);
-            }
-
-            let before = items.len();
-            items.retain(|item| item.task_id != task_id);
-            let removed = items.len() != before;
-
-            *queue = items.into_iter().collect();
-            removed
-        };
-
-        if removed {
-            if let Err(err) = self.persist_state().await {
-                warn!("Failed to persist state after dequeue: {}", err);
-            }
-        }
-        removed
-    }
-
-    async fn process_task_queue(&mut self) {
-        if self.queue_paused {
-            return;
-        }
-        loop {
-            self.settle_pending_semaphore_reduction();
-            self.reap_finished_active_downloads();
-            if self.active_downloads.len() >= self.config.concurrent_downloads {
-                break;
-            }
-            if self.download_semaphore.available_permits() == 0 {
-                break;
-            }
-
-            let next_task = {
-                let mut queue = self.task_queue.lock().await;
-                queue.pop()
-            };
-
-            let Some(task_priority) = next_task else {
-                break;
-            };
-
-            let task_id = task_priority.task_id.clone();
-            let task = match self.tasks.get(&task_id).cloned() {
-                Some(task) => task,
-                None => continue,
-            };
-
-            if task.status != TaskStatus::Pending
-                && task.status != TaskStatus::Paused
-                && task.status != TaskStatus::Failed
-            {
-                continue;
-            }
-
-            self.refresh_task_file_state(&task_id).await.ok();
-
-            let permit = match self.download_semaphore.clone().try_acquire_owned() {
-                Ok(permit) => permit,
-                Err(_) => {
-                    let _ = self.enqueue_task(&task_id, task_priority.priority).await;
-                    break;
-                }
-            };
-
-            if let Err(err) = self
-                .start_download_with_permit(&task_id, task, permit)
-                .await
-            {
-                warn!("Failed to start queued task {}: {}", task_id, err);
-            }
-        }
-    }
-
-    /// Apply deferred semaphore reductions when permits become available.
-    /// This keeps effective concurrency aligned with runtime config updates.
-    fn settle_pending_semaphore_reduction(&mut self) {
-        if self.pending_semaphore_reduction == 0 {
-            return;
-        }
-
-        let available = self.download_semaphore.available_permits();
-        if available == 0 {
-            return;
-        }
-
-        let to_forget = available.min(self.pending_semaphore_reduction);
-        let forgotten = self.download_semaphore.forget_permits(to_forget);
-
-        if forgotten > 0 {
-            self.pending_semaphore_reduction -= forgotten;
-            self.semaphore_capacity = self.semaphore_capacity.saturating_sub(forgotten);
-            debug!(
-                "Settled semaphore reduction: forgotten={}, pending={}, capacity={}",
-                forgotten, self.pending_semaphore_reduction, self.semaphore_capacity
-            );
-        }
-    }
-
     pub(crate) async fn scheduler_tick(&mut self) -> bool {
         if !self.is_running {
             return false;
@@ -2630,161 +2440,6 @@ impl DownloadManager {
             .map_err(|e| AppError::System(format!("Failed to persist completion marker: {}", e)))
     }
 
-    fn note_transfer_started(&mut self, task_id: &str) {
-        let should_reset = self
-            .task_lifecycle_timings
-            .get(task_id)
-            .and_then(|timing| timing.final_status.as_ref())
-            .is_some();
-
-        if should_reset || !self.task_lifecycle_timings.contains_key(task_id) {
-            self.task_lifecycle_timings.insert(
-                task_id.to_string(),
-                TaskLifecycleTiming {
-                    transfer_started_at: chrono::Utc::now(),
-                    commit_started_at: None,
-                    finished_at: None,
-                    final_status: None,
-                },
-            );
-        }
-    }
-
-    fn note_commit_started(&mut self, task_id: &str) {
-        let now = chrono::Utc::now();
-        let entry = self
-            .task_lifecycle_timings
-            .entry(task_id.to_string())
-            .or_insert(TaskLifecycleTiming {
-                transfer_started_at: now,
-                commit_started_at: None,
-                finished_at: None,
-                final_status: None,
-            });
-
-        if entry.commit_started_at.is_none() {
-            entry.commit_started_at = Some(now);
-        }
-        entry.final_status = None;
-        entry.finished_at = None;
-    }
-
-    fn note_terminal_status(&mut self, task_id: &str, final_status: &TaskStatus) {
-        let now = chrono::Utc::now();
-        let (transfer_started_at, commit_started_at) = {
-            let timing = self
-                .task_lifecycle_timings
-                .entry(task_id.to_string())
-                .or_insert(TaskLifecycleTiming {
-                    transfer_started_at: now,
-                    commit_started_at: None,
-                    finished_at: None,
-                    final_status: None,
-                });
-            (timing.transfer_started_at, timing.commit_started_at)
-        };
-
-        if let Some(commit_started_at) = commit_started_at {
-            let commit_duration = (now - commit_started_at)
-                .to_std()
-                .unwrap_or_default()
-                .as_secs_f64();
-
-            if *final_status == TaskStatus::Completed {
-                Self::push_metric_sample(
-                    &mut self.lifecycle_metrics.commit_duration_secs,
-                    commit_duration,
-                );
-
-                if commit_duration > 2.0 {
-                    self.lifecycle_metrics.commit_warning_count += 1;
-                    warn!(
-                        "Commit stage exceeded warning threshold for task {}: {:.3}s",
-                        task_id, commit_duration
-                    );
-                }
-                if commit_duration > 5.0 {
-                    self.lifecycle_metrics.commit_elevated_warning_count += 1;
-                    warn!(
-                        "Commit stage exceeded elevated threshold for task {}: {:.3}s",
-                        task_id, commit_duration
-                    );
-                }
-            } else if *final_status == TaskStatus::Failed {
-                self.lifecycle_metrics.failed_commit_count += 1;
-                warn!(
-                    "Task {} failed after entering commit stage; commit duration before failure {:.3}s",
-                    task_id, commit_duration
-                );
-            }
-        }
-
-        if *final_status == TaskStatus::Completed {
-            let transfer_end = commit_started_at.unwrap_or(now);
-            let transfer_duration = (transfer_end - transfer_started_at)
-                .to_std()
-                .unwrap_or_default()
-                .as_secs_f64();
-            let total_duration = (now - transfer_started_at)
-                .to_std()
-                .unwrap_or_default()
-                .as_secs_f64();
-
-            Self::push_metric_sample(
-                &mut self.lifecycle_metrics.transfer_duration_secs,
-                transfer_duration,
-            );
-            Self::push_metric_sample(
-                &mut self.lifecycle_metrics.total_duration_secs,
-                total_duration,
-            );
-        }
-
-        if let Some(timing) = self.task_lifecycle_timings.get_mut(task_id) {
-            timing.finished_at = Some(now);
-            timing.final_status = Some(final_status.clone());
-        }
-    }
-
-    fn note_peak_download_speed(&mut self, speed_bps: f64) {
-        if speed_bps.is_finite() && speed_bps > self.lifecycle_metrics.peak_download_speed_bps {
-            self.lifecycle_metrics.peak_download_speed_bps = speed_bps;
-        }
-    }
-
-    fn push_metric_sample(samples: &mut Vec<f64>, value: f64) {
-        if !value.is_finite() || value < 0.0 {
-            return;
-        }
-
-        samples.push(value);
-        const MAX_SAMPLES: usize = 512;
-        if samples.len() > MAX_SAMPLES {
-            let overflow = samples.len() - MAX_SAMPLES;
-            samples.drain(0..overflow);
-        }
-    }
-
-    fn average_metric(samples: &[f64]) -> f64 {
-        if samples.is_empty() {
-            0.0
-        } else {
-            samples.iter().sum::<f64>() / samples.len() as f64
-        }
-    }
-
-    fn percentile_metric(samples: &[f64], percentile: f64) -> f64 {
-        if samples.is_empty() {
-            return 0.0;
-        }
-
-        let mut values = samples.to_vec();
-        values.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
-        let clamped = percentile.clamp(0.0, 1.0);
-        let idx = ((values.len().saturating_sub(1)) as f64 * clamped).round() as usize;
-        values[idx]
-    }
-
     /// Update task status
     pub async fn update_task_status(&mut self, task_id: &str, status: TaskStatus) -> AppResult<()> {
         if let Some(task) = self.tasks.get_mut(task_id) {
@@ -2804,18 +2459,6 @@ impl DownloadManager {
             warn!("Failed to persist state after status update: {}", err);
         }
         Ok(())
-    }
-
-    /// Update download statistics
-    async fn update_stats(&mut self) {
-        self.recompute_stats();
-
-        // Emit stats updated event
-        if let Some(sender) = &self.event_sender {
-            let _ = sender.send(DownloadEvent::StatsUpdated {
-                stats: self.stats.clone(),
-            });
-        }
     }
 
     /// Apply state updates based on download events so backend state stays consistent with UI
@@ -3506,142 +3149,6 @@ impl DownloadManager {
                 Err(AppError::Download(error_msg))
             }
         }
-    }
-
-    /// Manually verify file integrity using specified algorithm
-    pub async fn verify_file_integrity(
-        &self,
-        file_path: &str,
-        algorithm: HashAlgorithm,
-    ) -> AppResult<IntegrityResult> {
-        info!(
-            "🔐 Starting manual integrity verification: {} with {:?}",
-            file_path, algorithm
-        );
-
-        if let Some(sender) = &self.event_sender {
-            let _ = sender.send(DownloadEvent::IntegrityCheckStarted {
-                task_id: "manual".to_string(),
-                algorithm: format!("{:?}", algorithm),
-            });
-        }
-
-        let result = self
-            .integrity_checker
-            .compute_hash(file_path, algorithm)
-            .await
-            .map_err(|e| AppError::System(format!("Integrity verification failed: {}", e)))?;
-
-        if let Some(sender) = &self.event_sender {
-            let _ = sender.send(DownloadEvent::IntegrityCheckCompleted {
-                task_id: "manual".to_string(),
-                result: result.clone(),
-            });
-        }
-
-        info!(
-            "✅ Manual integrity verification completed: {} - Valid: {}",
-            file_path, result.is_valid
-        );
-        Ok(result)
-    }
-
-    /// Verify multiple files with different algorithms concurrently
-    pub async fn verify_batch_integrity(
-        &self,
-        files: Vec<(String, HashAlgorithm)>,
-    ) -> AppResult<Vec<(String, IntegrityResult)>> {
-        info!(
-            "🔐 Starting batch integrity verification for {} files",
-            files.len()
-        );
-
-        // For now, verify files sequentially until batch method is implemented
-        let mut results = Vec::new();
-        for (file_path, algorithm) in files {
-            match self
-                .integrity_checker
-                .compute_hash(&file_path, algorithm)
-                .await
-            {
-                Ok(result) => results.push((file_path, result)),
-                Err(e) => {
-                    warn!("Failed to verify {}: {}", file_path, e);
-                    // Continue with other files
-                }
-            }
-        }
-
-        info!(
-            "✅ Batch integrity verification completed: {} files",
-            results.len()
-        );
-        Ok(results)
-    }
-
-    /// Compute hash of file without comparison (for getting file hashes)
-    pub async fn compute_file_hash(
-        &self,
-        file_path: &str,
-        algorithm: HashAlgorithm,
-    ) -> AppResult<String> {
-        info!("🧮 Computing hash for: {} with {:?}", file_path, algorithm);
-
-        let result = self
-            .integrity_checker
-            .compute_hash(file_path, algorithm)
-            .await
-            .map_err(|e| AppError::System(format!("Hash computation failed: {}", e)))?;
-
-        info!("✅ Hash computed: {} - {}", file_path, result.computed_hash);
-        Ok(result.computed_hash)
-    }
-
-    /// Set expected hash for a URL to enable automatic verification
-    pub async fn set_expected_hash(&mut self, url: &str, hash: &str) -> AppResult<()> {
-        self.config
-            .expected_hashes
-            .insert(url.to_string(), hash.to_string());
-        info!("🎯 Set expected hash for {}: {}", url, hash);
-        Ok(())
-    }
-
-    /// Remove expected hash for a URL
-    pub async fn remove_expected_hash(&mut self, url: &str) -> AppResult<()> {
-        self.config.expected_hashes.remove(url);
-        info!("🗑️ Removed expected hash for: {}", url);
-        Ok(())
-    }
-
-    /// Get all expected hashes
-    pub fn get_expected_hashes(&self) -> &HashMap<String, String> {
-        &self.config.expected_hashes
-    }
-
-    /// Enable or disable automatic integrity verification
-    pub async fn set_auto_integrity_verification(&mut self, enabled: bool) -> AppResult<()> {
-        self.config.auto_verify_integrity = enabled;
-        info!(
-            "🔧 Auto integrity verification: {}",
-            if enabled { "enabled" } else { "disabled" }
-        );
-        Ok(())
-    }
-
-    /// Set the default integrity algorithm
-    pub async fn set_integrity_algorithm(&mut self, algorithm: HashAlgorithm) -> AppResult<()> {
-        let algorithm_str = match algorithm {
-            HashAlgorithm::Sha256 => "sha256",
-            HashAlgorithm::Sha512 => "sha512",
-            HashAlgorithm::Blake2b512 => "blake2b512",
-            HashAlgorithm::Blake2s256 => "blake2s256",
-            HashAlgorithm::Md5 => "md5",
-            HashAlgorithm::Sha1 => "sha1",
-        };
-
-        self.config.integrity_algorithm = Some(algorithm_str.to_string());
-        info!("🔧 Default integrity algorithm set to: {:?}", algorithm);
-        Ok(())
     }
 
     /// Get retry statistics from the retry executor
