+++ src-tauri/src/core/manager/events.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+impl DownloadManager {
+    pub(super) fn emit_event(&self, event: DownloadEvent) {
+        if let Some(sender) = &self.event_sender {
+            let _ = sender.send(event);
+        }
+    }
+}
